# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Product Sustainability & Impact Assessment API",
    "description": "Assess product sustainability and environmental impact (CO2, water) for e-commerce and AI agents. Returns explainable, structured scores with human-readable explanations.",
    "version": "1.2.0"
  },
  "paths": {
    "/v1/methodology": {
      "get": {
        "summary": "Get calculation methodology",
        "description": "Returns a JSON explanation of how sustainability scores and impact estimates (CO2, water, breakdown) are calculated. Use this to build trust and explain results to users or AI agents.",
        "operationId": "get_methodology_v1_methodology_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MethodologyResponse"
                }
              }
            }
          }
        }
      }
    },
    "/v1/assess-impact": {
      "post": {
        "summary": "Assess product sustainability and impact",
        "description": "Submit product details to receive an enterprise-grade response: score, confidence_level, CO2, explainable breakdown, CBAM analysis, and human-readable explanations. Rate limited per API key.",
        "operationId": "post_assess_impact_v1_assess_impact_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AssessImpactRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ImpactAssessmentResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "APIKeyHeader": []
          }
        ]
      }
    },
    "/demo/score": {
      "get": {
        "tags": [
          "Demo"
        ],
        "summary": "Demo: See a perfect score example",
        "description": "Returns a hardcoded, high-scoring example so users can instantly see what the API is capable of—no API key required.",
        "operationId": "demo_score_demo_score_get",
        "parameters": [
          {
            "name": "product_name",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Product Name"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ImpactAssessmentResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/health": {
      "get": {
        "tags": [
          "Health"
        ],
        "summary": "Health",
        "description": "Health check for load balancers and monitoring.",
        "operationId": "health_health_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "AssessImpactRequest": {
        "properties": {
          "product_name": {
            "type": "string",
            "title": "Product Name",
            "description": "Human-readable name of the product (e.g. 'Organic Cotton T-Shirt'). Echoed in response."
          },
          "material_composition": {
            "additionalProperties": {
              "type": "number"
            },
            "type": "object",
            "title": "Material Composition",
            "description": "Share of each material by weight, as decimals summing to 1.0. Example: {\"polyester\": 0.6, \"cotton\": 0.4}."
          },
          "weight_kg": {
            "type": "number",
            "exclusiveMinimum": 0.0,
            "title": "Weight Kg",
            "description": "Product weight in kilograms. Must be greater than 0."
          },
          "origin_country": {
            "type": "string",
            "title": "Origin Country",
            "description": "ISO-style country code or name where the product is produced (e.g. 'CN', 'Bangladesh'). Affects logistics emissions."
          },
          "destination_country": {
            "type": "string",
            "title": "Destination Country",
            "description": "ISO-style country code or name where the product is delivered (e.g. 'US', 'Germany'). Affects logistics emissions."
          },
          "shipping_mode": {
            "type": "string",
            "title": "Shipping Mode",
            "description": "Shipping mode: 'sea', 'air', 'road', or 'rail'. Affects CO2 calculation with circuitry factors and mode multipliers.",
            "default": "sea"
          }
        },
        "type": "object",
        "required": [
          "product_name",
          "material_composition",
          "weight_kg",
          "origin_country",
          "destination_country"
        ],
        "title": "AssessImpactRequest",
        "description": "Request body for the impact assessment endpoint."
      },
      "CbamAnalysis": {
        "properties": {
          "is_relevant": {
            "type": "boolean",
            "title": "Is Relevant",
            "description": "True if the product contains CBAM-relevant materials (steel, aluminum, cement, fertilizer, hydrogen, iron)."
          },
          "reason": {
            "type": "string",
            "title": "Reason",
            "description": "Short explanation: either which materials triggered CBAM relevance or that none were found."
          }
        },
        "type": "object",
        "required": [
          "is_relevant",
          "reason"
        ],
        "title": "CbamAnalysis",
        "description": "CBAM relevance and human-readable reason for regulatory awareness."
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "ImpactAssessmentResponse": {
        "properties": {
          "product_name": {
            "type": "string",
            "title": "Product Name",
            "description": "Product name as submitted; echoed for context."
          },
          "total_sustainability_score": {
            "type": "number",
            "maximum": 100.0,
            "minimum": 0.0,
            "title": "Total Sustainability Score",
            "description": "Overall sustainability score from 0 to 100. Higher is better. Score = 0.5*Material + 0.3*Logistics + 0.2*Weight."
          },
          "confidence_level": {
            "type": "string",
            "title": "Confidence Level",
            "description": "Confidence in the estimate (e.g. low, medium, high). Currently medium for model-based estimates.",
            "default": "medium"
          },
          "co2_estimate_kg": {
            "type": "number",
            "minimum": 0.0,
            "title": "Co2 Estimate Kg",
            "description": "Estimated CO2 equivalent in kg for materials and logistics. Indicative, not certified."
          },
          "breakdown": {
            "$ref": "#/components/schemas/ImpactBreakdown",
            "description": "Per-dimension scores so agents can explain e.g. 'The score is low because logistics_score is low.'"
          },
          "cbam_analysis": {
            "$ref": "#/components/schemas/CbamAnalysis",
            "description": "Whether the product is CBAM-relevant and why; key for European enterprise buyers."
          },
          "explanation": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Explanation",
            "description": "Human-readable explanations for the score. AI agents can surface these to users to explain why the score is high or low."
          },
          "methodology_version": {
            "type": "string",
            "title": "Methodology Version",
            "description": "Version of the calculation methodology. Use to track changes and build trust.",
            "default": "v1.2.0"
          },
          "disclaimer": {
            "type": "string",
            "title": "Disclaimer",
            "description": "Legal disclaimer: indicative model-based estimate; not for regulatory CBAM filings."
          }
        },
        "type": "object",
        "required": [
          "product_name",
          "total_sustainability_score",
          "co2_estimate_kg",
          "breakdown",
          "cbam_analysis",
          "disclaimer"
        ],
        "title": "ImpactAssessmentResponse",
        "description": "Enterprise-grade, nested, explainable response for AI agents and e-commerce."
      },
      "ImpactBreakdown": {
        "properties": {
          "material_score": {
            "type": "number",
            "title": "Material Score",
            "description": "Sustainability score for materials only (0–100). Higher = lower impact materials."
          },
          "logistics_score": {
            "type": "number",
            "title": "Logistics Score",
            "description": "Sustainability score for shipping from origin to destination (0–100). Higher = lower logistics impact."
          },
          "weight_penalty": {
            "type": "number",
            "title": "Weight Penalty",
            "description": "Penalty from product weight; heavier products add to this. Used in total score calculation."
          }
        },
        "type": "object",
        "required": [
          "material_score",
          "logistics_score",
          "weight_penalty"
        ],
        "title": "ImpactBreakdown",
        "description": "Per-dimension scores so agents can explain why a score is low or high."
      },
      "MethodologyResponse": {
        "properties": {
          "methodology_version": {
            "type": "string",
            "title": "Methodology Version",
            "description": "Current methodology version used by the API.",
            "default": "v1.2.0"
          },
          "description": {
            "type": "string",
            "title": "Description",
            "description": "Short summary of the methodology purpose and scope."
          },
          "total_sustainability_score": {
            "additionalProperties": true,
            "type": "object",
            "title": "Total Sustainability Score",
            "description": "How total_sustainability_score is computed (formula and weights)."
          },
          "co2_estimate_kg": {
            "additionalProperties": true,
            "type": "object",
            "title": "Co2 Estimate Kg",
            "description": "How CO2 equivalent is estimated (materials + logistics)."
          },
          "water_usage_liters": {
            "additionalProperties": true,
            "type": "object",
            "title": "Water Usage Liters",
            "description": "How water usage is estimated (material-based)."
          },
          "breakdown": {
            "additionalProperties": true,
            "type": "object",
            "title": "Breakdown",
            "description": "How material_score, logistics_score, and weight_penalty are derived."
          },
          "disclaimer": {
            "type": "string",
            "title": "Disclaimer",
            "description": "Legal/disclaimer text: results are indicative, not certified LCA."
          }
        },
        "type": "object",
        "required": [
          "description",
          "total_sustainability_score",
          "co2_estimate_kg",
          "water_usage_liters",
          "breakdown",
          "disclaimer"
        ],
        "title": "MethodologyResponse",
        "description": "Static JSON explaining how scores are calculated (for transparency and trust)."
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "APIKeyHeader": {
        "type": "apiKey",
        "in": "header",
        "name": "X-API-Key",
        "description": "API key for authentication. Get your key from the provider."
      }
    }
  },
  "security": [
    {
      "APIKeyHeader": []
    }
  ]
}